+++ src/Service/Indexing/AbstractIndexingElastic.php
@@ -4,12 +4,7 @@ namespace App\Service\Indexing;
 
 use App\Exception\IndexingException;
 use App\Model\Indexing\IndexNames;
-use App\Model\Indexing\Mappings\EventWithOccurrences;
-use App\Model\Indexing\Mappings\Location;
-use App\Model\Indexing\Mappings\OccurrenceWithEvent;
-use App\Model\Indexing\Mappings\Organizer;
-use App\Model\Indexing\Mappings\Tag;
-use App\Model\Indexing\Mappings\Vocabularies;
+use App\Model\Indexing\Mappings\MappingsProvider;
 use Elastic\Elasticsearch\Client;
 use Elastic\Elasticsearch\Exception\ClientResponseException;
 use Elastic\Elasticsearch\Exception\MissingParameterException;
@@ -354,10 +349,7 @@ abstract class AbstractIndexingElastic implements IndexingInterface
                             ],
                         ],
                     ],
-                    'mappings' => [
-                        'dynamic' => 'strict',
-                        'properties' => $this->getIndexProperties(),
-                    ],
+                    'mappings' => MappingsProvider::mappingFor(IndexNames::from($this::INDEX_ALIAS)),
                 ],
             ]);
 
@@ -373,23 +365,5 @@ abstract class AbstractIndexingElastic implements IndexingInterface
     public function criteria(): array
     {
         return [];
-    }
-
-    /**
-     * @throws IndexingException
-     */
-    private function getIndexProperties(): array
-    {
-        $index = IndexNames::from($this::INDEX_ALIAS);
-
-        return match ($index) {
-            IndexNames::Organizations => Organizer::getProperties(),
-            IndexNames::Events => EventWithOccurrences::getProperties(),
-            IndexNames::Locations => Location::getProperties(),
-            IndexNames::Tags => Tag::getProperties(),
-            IndexNames::Vocabularies => Vocabularies::getProperties(),
-            IndexNames::Occurrences, IndexNames::DailyOccurrences => OccurrenceWithEvent::getProperties(),
-            // IndexNames::ApiKeys => throw new \Exception('To be implemented'),
-        };
     }
 }